Chain numbers and letters and jumps, oh my!

In Rosetta there's several ways of specifying a "chain": chain numbers (chain ID), chain letters, and jump numbers. These are not interchangeable. In general, there is *NOT* a 1:1 correspondence between chain numbers, chain letters or jump numbers, unless the pose has been specifically set up as such.

To help reduce confusion, add documentation to (new function) Pose::num_chains() which describes the various options, and explicitly points out they may be different. Also add pointers to that documentation from various places where the distinction is relevant, or where a naive person might find a function that works with "chains" and might need clarification as to the distinction.

Fix a number of utility functions in core/pose/util.hh which work with chain numbers/letters/jumps so that they're more robust to poses where the correspondence is not 1:1, and unit test them with a funky pose.

Bonus: I've added C++11 move constructors and initialization list constructors to utility::vector1/0/L, so creating vector1s with defined contents is easy: 

    utility::vector1<core::Size> single_digit_primes{2,3,5,7};

Old repository SHA1: 8a96f86dc7ef2da0a7f81d790e3e7c1b968638bb

+++ source/src/devel/domain_insertion/FusePosesNtoCMover.cc
@@ -856,7 +856,7 @@ SetupCoiledCoilFoldTreeMover::apply( core::pose::Pose & pose )
 			std::cout << "MEEP" << std::endl;
 			new_ft.add_edge( e.start(), cut_seqpos, e.label() );
 			new_ft.add_edge( e.stop(), cut_seqpos + 1, e.label() );
-		} else if ( (e.is_jump() ) && ( pose.chain( e.start() ) == (int) pose_chains[1]) && ( pose.chain( e.stop() ) == (int) pose_chains[2] ) ) {
+		} else if ( (e.is_jump() ) && ( pose.chain( e.start() ) == pose_chains[1]) && ( pose.chain( e.stop() ) == pose_chains[2] ) ) {
 			relevant_jump_found = true;
 			new_ft.add_edge( pose.conformation().chain_begin( pose_chains[1] ), pose.conformation().chain_begin( pose_chains[2] ), e.label() );
 			new_ft.add_edge( pose.conformation().chain_end( pose_chains[1] ), pose.conformation().chain_end( pose_chains[2] ), jump_count );
